# docs/swagger.yaml
swagger: "2.0"
info:
  description: |-
     User Management System
     HTTP Status:
     200 (Success)
     203 (Error)
  version: 0.0.1
  title: User Management System
schemes:
  - http
paths:
  /users/api/v1/get/users:
    get:
      summary: api to get users list
      operationId: getUsers
      tags:
        - Users
      consumes:
        - application/json
      parameters:
        - $ref: "#/components/parameters/headerPlatform"
        - $ref: "#/components/parameters/headerLanguage"
        - in: query
          name: user_id
          type: number
          description: please enter user_id
        - in: query
          name: page
          type: number
          required: true
          description: page number minimum value will be 1
        - in: query
          name: limit
          type: number
          description: record per page you needed it is optional default set 10.
        - in: query
          name: text
          type: string
          description: search by owner name optional field.
      responses:
        "200":
          description: get user details in response
          schema:
            allOf:
              - $ref: "#/components/schemas/ApiResponse"
        "400":
          $ref: "#/components/responses/BadRequestError"
...
components:
...
  parameters:
    headerLanguage:
      name: language
      in: header
      required: true
      schema:
        type: string
        default: en
        enum:
          - en
    headerPlatform:
      name: os
      in: header
      required: true
      schema:
        type: string
        default: android
        enum:
          - ios
          - android
...
  schemas:
    ApiResponse:
      required:
        - success
        - message
      properties:
        success:
          type: boolean
        message:
          type: string
        data: {}
  responses:
    BadRequestError:
      description: Request fails validation or doesn't meet all conditions for request
      content:
        application/json:
          schema:
            example:
              success: false
              message: Validation failed
              data: ""
            allOf:
              - $ref: "#/components/schemas/ApiResponse"
              - type: object
                properties:
                  data:
                    type: object
                    nullable: true
...